+++ packages/@expo/metro-config/src/transform-worker/asset-transformer.ts
@@ -8,8 +8,7 @@
  * Fork of the upstream transformer, but with modifications made for web production hashing.
  * https://github.com/facebook/metro/blob/412771475c540b6f85d75d9dcd5a39a6e0753582/packages/metro-transform-worker/src/utils/assetTransformer.js#L1
  */
-import template from '@babel/template';
-import * as t from '@babel/types';
+import { template, types as t } from '@babel/core';
 import { generateAssetCodeFileAst } from 'metro/src/Bundler/util';
 import { BabelTransformerArgs } from 'metro-babel-transformer';
 import path from 'node:path';