+++ packages/babel-preset-expo/src/server-actions-plugin.ts
@@ -8,16 +8,7 @@
  * https://github.com/lubieowoce/tangle/blob/5229666fb317d0da9363363fc46dc542ba51e4f7/packages/babel-rsc/src/babel-rsc-actions.ts#L1C1-L909C25
  */
 
-import {
-  ConfigAPI,
-  types,
-  template,
-  type NodePath,
-  type PluginObj,
-  type PluginPass,
-} from '@babel/core';
-import type { Scope as BabelScope } from '@babel/traverse';
-import * as t from '@babel/types';
+import type { ConfigAPI, types as t, NodePath, PluginObj, PluginPass } from '@babel/core';
 import { relative as getRelativePath } from 'node:path';
 import url from 'node:url';
 
@@ -34,24 +25,26 @@ type ExtractedActionInfo = { localName?: string; exportedName: string };
 
 const LAZY_WRAPPER_VALUE_KEY = 'value';
 
-// React doesn't like non-enumerable properties on serialized objects (see `isSimpleObject`),
-// so we have to use closure scope for the cache (instead of a non-enumerable `this._cache`)
-const _buildLazyWrapperHelper = template(`(thunk) => {
-  let cache;
-  return {
-    get ${LAZY_WRAPPER_VALUE_KEY}() {
-      return cache || (cache = thunk());
-    }
-  }
-}`);
-
-const buildLazyWrapperHelper = () => {
-  return (_buildLazyWrapperHelper() as t.ExpressionStatement).expression;
-};
-
 export function reactServerActionsPlugin(
-  api: ConfigAPI & { types: typeof types }
+  api: ConfigAPI & typeof import('@babel/core')
 ): PluginObj<PluginPass> {
+  const { types: t } = api;
+
+  // React doesn't like non-enumerable properties on serialized objects (see `isSimpleObject`),
+  // so we have to use closure scope for the cache (instead of a non-enumerable `this._cache`)
+  const _buildLazyWrapperHelper = api.template(`(thunk) => {
+    let cache;
+    return {
+      get ${LAZY_WRAPPER_VALUE_KEY}() {
+        return cache || (cache = thunk());
+      }
+    }
+  }`);
+
+  const buildLazyWrapperHelper = () => {
+    return (_buildLazyWrapperHelper() as t.ExpressionStatement).expression;
+  };
+
   const possibleProjectRoot = api.caller(getPossibleProjectRoot);
   let addReactImport: () => t.Identifier;
   let wrapBoundArgs: (expr: t.Expression) => t.Expression;
@@ -133,7 +126,7 @@ export function reactServerActionsPlugin(
     const isPathFunctionInTopLevel = path.find((p) => p.isProgram()) === path;
 
     const decl = isPathFunctionInTopLevel ? path : findImmediatelyEnclosingDeclaration(path);
-    let inserted: NodePath<types.ExportNamedDeclaration>;
+    let inserted: NodePath<t.ExportNamedDeclaration>;
 
     const canInsertExportNextToPath = (decl: NodePath) => {
       if (!decl) {
@@ -186,9 +179,10 @@ export function reactServerActionsPlugin(
     } else {
       // Fallback to inserting after the last import if no enclosing declaration is found
       const programBody = moduleScope.path.get('body');
-      const lastImportPath = findLast(
-        Array.isArray(programBody) ? programBody : [programBody],
-        (stmt) => stmt.isImportDeclaration()
+      const lastImportPath = (Array.isArray(programBody) ? programBody : [programBody]).findLast(
+        (statement) => {
+          return statement.isImportDeclaration();
+        }
       );
 
       [inserted] = lastImportPath!.insertAfter(functionDeclaration);
@@ -234,6 +228,11 @@ export function reactServerActionsPlugin(
       boundArgs,
     ]);
   };
+
+  function hasUseServerDirective(path: FnPath) {
+    const { body } = path.node;
+    return t.isBlockStatement(body) && body.directives.some((d) => d.value.value === 'use server');
+  }
 
   return {
     name: 'expo-server-actions',
@@ -718,9 +717,9 @@ const isChildScope = ({
   parent,
   child,
 }: {
-  root: BabelScope;
-  parent: BabelScope;
-  child: BabelScope;
+  root: NodePath['scope'];
+  parent: NodePath['scope'];
+  child: NodePath['scope'];
 }) => {
   let curScope = child;
   while (curScope !== root) {
@@ -730,13 +729,6 @@ const isChildScope = ({
     curScope = curScope.parent;
   }
   return false;
-};
-
-const findLast = <T>(arr: T[], predicate: (value: T) => boolean): T | undefined => {
-  for (let i = arr.length - 1; i >= 0; i--) {
-    if (predicate(arr[i])) return arr[i];
-  }
-  return undefined;
 };
 
 function findImmediatelyEnclosingDeclaration(path: FnPath) {
@@ -795,9 +787,4 @@ function assertExpoMetadata(metadata: any): asserts metadata is {
   if (!metadata || typeof metadata !== 'object') {
     throw new Error('Expected Babel state.file.metadata to be an object');
   }
-}
-
-function hasUseServerDirective(path: FnPath) {
-  const { body } = path.node;
-  return t.isBlockStatement(body) && body.directives.some((d) => d.value.value === 'use server');
 }